change in migration, added logic for fetching score board and grades report of students, adding validation in model and action for fetching score board

+++ db/migrate/20231208040831_create_students.rb
@@ -1,11 +1,11 @@
 class CreateStudents < ActiveRecord::Migration[6.1]
   def change
     create_table :students do |t|
-      t.float :subject_1
-      t.float :subject_2
-      t.float :subject_3
-      t.float :subject_4
-      t.float :total_score
+      t.integer :subject_1
+      t.integer :subject_2
+      t.integer :subject_3
+      t.integer :subject_4
+      t.integer :total_score
       t.boolean :above_average, default: false
       t.string :grade
       t.timestamps
